pyproject.toml
[project]
name = "pipecat-quickstart"
version = "0.1.0"
description = "Quickstart example for building voice AI bots with Pipecat"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[webrtc,silero,soniox,azure,elevenlabs,local-smart-turn-v3,runner]",
    "pipecat-ai-cli",
    "pyyaml>=6.0",
    "upstash-redis>=0.15.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.1",
    "google-api-python-client>=2.187.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
